requirements.txt
gradio
transformers
torch
Pillow
huggingface_hub